chore: Add backoffice/clients portal and Aristos view
The code changes include adding a backoffice/clients portal and a new view called Aristos. This is in line with the current project requirements mentioned in the README file. The web.php file has been modified to include a new route for the Aristos view. The Aristos view itself has been added with a simple "Hello Aristos" message.

Please remove any meta information and format the commit message according to the repository's guidelines.